core: serialize RetAdjusted.pop_bytes as u64 to match the IR field
clang -Werror -Wconversion rejected the implicit u64->u32 narrowing in the
modern serializer; the format is unreleased, so writer and reader move to
the full-width field instead of truncating.

+++ core/src/ir/serialization.cpp
@@ -323,7 +323,7 @@ private:
                 u64(x.return_guest_pc);
             } else if constexpr (std::is_same_v<T, RetAdjusted>) {
                 enum8(OpTag::RetAdjusted);
-                u32(x.pop_bytes);
+                u64(x.pop_bytes);
             } else if constexpr (std::is_same_v<T, Cpuid>) {
                 enum8(OpTag::Cpuid);
             } else if constexpr (std::is_same_v<T, Syscall>) {
@@ -943,8 +943,8 @@ read_optional_ref(Reader& reader) {
             return Op{CallReg{target, return_guest_pc}};
         }
         case OpTag::RetAdjusted: {
-            std::uint32_t pop_bytes = 0;
-            if (!reader.u32(pop_bytes)) return reader.error();
+            std::uint64_t pop_bytes = 0;
+            if (!reader.u64(pop_bytes)) return reader.error();
             return Op{RetAdjusted{pop_bytes}};
         }
         case OpTag::Cpuid:
